chore: quality of code

+++ src/encoding/xor/xor.h
@@ -3,6 +3,6 @@
 
 #include <linux/types.h>
 
-int enc_xor_process(u8 index, const u8 *req_buf, size_t req_buflen, u8 **res_buf, size_t *res_buflen);
+int enc_xor_process(size_t index, const u8 *req_buf, size_t req_buflen, u8 **res_buf, size_t *res_buflen);
 
 #endif
